requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
joblib>=1.3
pyyaml>=6.0
mlflow>=2.12
tensorflow>=2.16
httpx>=0.27
orjson>=3.10
